Switch to having the `BoxedAccount` struct in azalea-client, this makes it so that azalea-auth doesnt need to depend on the ecs anymore
