+++ azalea-auth/Cargo.toml
@@ -27,7 +27,6 @@ thiserror = "1.0.57"
 tokio = { version = "1.36.0", features = ["fs"] }
 uuid = { version = "1.7.0", features = ["serde", "v3"] }
 md-5 = "0.10.6"
-bevy_ecs = "0.13.0"
 async-trait = "0.1.78"
 
 [dev-dependencies]
