+++ azalea-client/src/lib.rs
@@ -30,7 +30,7 @@ pub mod task_pool;
 
 pub use azalea_protocol::packets::configuration::serverbound_client_information_packet::ClientInformation;
 pub use client::{
-    start_ecs_runner, Client, DefaultPlugins, JoinError, JoinedClientBundle, TickBroadcast,
+    start_ecs_runner, Client, DefaultPlugins, JoinError, JoinedClientBundle, TickBroadcast, BoxedAccount,
 };
 pub use events::Event;
 pub use local_player::{GameProfileComponent, InstanceHolder, TabList};
